Added a method to calculate the buy maximum.
Wrote a quick binary search algorithm to find the maximum number of the selected shop item that the player can buy based on available inventory space and current gold. Decided to do it this way to avoid having to implement multi-byte division and to see how it would feel to hand compile a JS program into 6502. It was super fun!